refactor(Calcified): use `Int`s instead of `Byte`s for port numbers i…

+++ src/main/kotlin/dev/frozenmilk/dairy/calcified/hardware/encoder/TicksEncoder.kt
@@ -3,7 +3,7 @@ package dev.frozenmilk.dairy.calcified.hardware.encoder
 import dev.frozenmilk.dairy.calcified.hardware.CalcifiedModule
 import dev.frozenmilk.dairy.core.util.supplier.numeric.EnhancedDoubleSupplier
 
-class TicksEncoder internal constructor(module: CalcifiedModule, port: Byte) : CalcifiedEncoder<Double>(module, port) {
+class TicksEncoder internal constructor(module: CalcifiedModule, port: Int) : CalcifiedEncoder<Double>(module, port) {
 	override val enhancedSupplier = EnhancedDoubleSupplier({ module.bulkData.getEncoder(port.toInt()).toDouble() * direction.multiplier })
 	override val zero = 0.0
 }
